Move widgets under named-packages
