+++ AppWidget/app/src/main/AndroidManifest.xml
@@ -1,5 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
+<?xml version="1.0" encoding="utf-8"?><!--
   ~ Copyright (C) 2021 The Android Open Source Project
   ~
   ~ Licensed under the Apache License, Version 2.0 (the "License");
@@ -26,7 +25,16 @@
         android:theme="@style/Theme.AppWidget">
 
         <activity
-            android:name=".ListWidgetConfigureActivity"
+            android:name=".MainActivity"
+            android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+
+        <activity
+            android:name=".list.ListWidgetConfigureActivity"
             android:exported="false">
             <intent-filter>
                 <action android:name="android.appwidget.action.APPWIDGET_CONFIGURE" />
@@ -34,7 +42,7 @@
         </activity>
 
         <receiver
-            android:name=".ListAppWidget"
+            android:name=".list.ListAppWidget"
             android:exported="false">
             <intent-filter>
                 <action android:name="android.appwidget.action.APPWIDGET_UPDATE" />
@@ -46,7 +54,7 @@
         </receiver>
 
         <receiver
-            android:name=".ItemsCollectionAppWidget"
+            android:name=".buttons.ItemsCollectionAppWidget"
             android:exported="false">
             <intent-filter>
                 <action android:name="android.appwidget.action.APPWIDGET_UPDATE" />
@@ -58,7 +66,7 @@
         </receiver>
 
         <receiver
-            android:name=".WeatherForecastAppWidget"
+            android:name=".weather.WeatherForecastAppWidget"
             android:exported="false">
             <intent-filter>
                 <action android:name="android.appwidget.action.APPWIDGET_UPDATE" />
@@ -68,19 +76,6 @@
                 android:name="android.appwidget.provider"
                 android:resource="@xml/app_widget_weather_forecast" />
         </receiver>
-
-        <service
-            android:name=".ItemsCollectionRemoteViewsService"
-            android:permission="android.permission.BIND_REMOTEVIEWS" />
-
-        <activity
-            android:name=".MainActivity"
-            android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
     </application>
 
 </manifest>
